requirements.txt
pandas>=2.0.0
scikit-learn>=1.2.0
mlxtend>=0.23.0
streamlit>=1.30.0
matplotlib>=3.7.0
seaborn>=0.12.0
shap>=0.44.0
joblib>=1.3.0
openpyxl>=3.1.0
requests>=2.30.0
umap-learn>=0.5.5
tensorflow>=2.15.0
